Don't read MCG_CTL if it's not present
rdmsr with MCG_CTL in ecx will result a GP# if it is simply not available.
The msr driver in kernel is now clever enough to deal with this problem
with an exception table, but it is simply not necessary to try.

lsmsr has the same prblem as x86info, but I can't figure out a nice
workaround.

Signed-off-by: Jike Song <albcamus@gmail.com>