Enhance descriptions for Intel feature flags.
I have added descriptions for more of the Intel flags.  Also added a
couple missing flags which are documented in the SDM section on CPUID
function.

Modified the flag_decode() function, so that it can print the flags.
It probably could be done in a more elegant way.

I also changed the verbose output so that it prints:
    "short_name description" instead of just description.