cpuid.c: Use native cpuid call silently instead of failing
x86info cpuid.c: Use native cpuid call silently instead of failing

If cpuid driver is not loaded a warning is printed while the output
should still be correct.

Use the cpuid assembler call silently instead of printing:
WARNING: Detected SMP, but unable to access cpuid driver.
Used Uniprocessor CPU routines. Results inaccurate.

For that some sanity checks are added to native_cpuid and it now returns an
error on failure.

Also the sched affinity mask is restored which is an additional cleanup.
Not sure whether it's worth it, but later code will only get executed on a
specific cpu instead letting the scheduler decide which looks wrong.

Signed-off-by: Thomas Renninger <trenn@suse.de>
CC: davej@redhat.com