IA32_PERF_STATUS is model specific.
We shouldn't try to parse it using the encoding of one ancient CPU,
as the results are likely to be incorrect.

Just dump it raw for now.  I don't particularly feel like looking
up the encoding for every CPU, so this is probably how this is going
to stay. For people who care, they can download the spec for their
CPU and decode it by hand.