x86info, AMD: Update CPU revisions
Provide just a general model description for all family 10h CPUs. It's
not possible to derive from model/family/stepping whether its an
Phenom or Turion II etc. This information is provided in the processor
name string. (It could be decoded from other CPUID information and
with help of CPU revision guides but that's not yet implemented.)

Signed-off-by: Andreas Herrmann <andreas.herrmann3@amd.com>