Add exact steppings and complete list of all bloomfield cores.
The information is taken from page 19 of the document 320836-036US
"Intel CoreTM i7-900 Desktop Processor Extreme Edition Series and Intel
CoreTM i7-900 Desktop Processor Series Specification Update".
Tested on SLBKP (2.8Ghz variant).  This should be a complete list of
bloomfield CPUs.

The change removes the only use of the ONLYMHZ flag, but I kept the flag
and the supporting code, since it might be useful in other situations.